updated services, schema, og_panels, image, css_gzip, and cacherouter

+++ cacherouter/engines/memcache.php
@@ -1,4 +1,10 @@
 <?php
+/**
+ * $Id: memcache.php,v 1.1.2.13 2009/09/05 13:03:25 slantview Exp $
+ *
+ * @file memcache.php
+ *   Engine file for memcache.
+ */
 class memcacheCache extends Cache {
   var $settings = array();
   var $memcache;
@@ -7,21 +13,18 @@ class memcacheCache extends Cache {
     return $this->fast_cache;
   }
   
-  function __construct($bin) {
-    global $conf;
+  function __construct($bin, $options, $default_options) {
     // Assign the servers on the following order: bin specific -> default specific -> localhost port 11211
-    if (isset($conf['cacherouter'][$bin]['server'])) {
-    	$this->settings['servers'] = $conf['cacherouter'][$bin]['server'];
-    	$this->settings['compress'] = isset($conf['cacherouter'][$bin]['compress']) ? MEMCACHE_COMPRESSED : 0;
-      $this->settings['shared'] = isset($conf['cacherouter'][$bin]['shared']) ?
-                                  $conf['cacherouter'][$bin]['shared'] : TRUE;
+    if (isset($options['servers'])) {
+    	$this->settings['servers'] = $options['servers'];
+    	$this->settings['compress'] = isset($options['compress']) ? MEMCACHE_COMPRESSED : 0;
+      $this->settings['shared'] = isset($options['shared']) ? $options['shared'] : TRUE;
     }
     else {
-      if (isset($conf['cacherouter']['default']['server'])) {
-        $this->settings['servers'] = $conf['cacherouter']['default']['server'];
-        $this->settings['compress'] = isset($conf['cacherouter']['default']['compress']) ? MEMCACHE_COMPRESSED : 0;
-        $this->settings['shared'] = isset($conf['cacherouter']['default']['shared']) ?
-                                    $conf['cacherouter']['default']['shared'] : TRUE;
+      if (isset($default_options['servers'])) {
+        $this->settings['servers'] = $default_options['servers'];
+        $this->settings['compress'] = isset($default_options['compress']) ? MEMCACHE_COMPRESSED : 0;
+        $this->settings['shared'] = isset($default_options['shared']) ? $default_options['shared'] : TRUE;
       }
       else {
         $this->settings['servers'] = array('localhost:11211');
@@ -30,7 +33,7 @@ class memcacheCache extends Cache {
       }
     }
                                 
-    parent::__construct($bin);
+    parent::__construct($bin, $options, $default_options);
     
     $this->connect();
   }
@@ -51,11 +54,7 @@ class memcacheCache extends Cache {
     return $cache;
   }
   
-  function set($key, $value, $expire = CACHE_PERMANENT, $headers = NULL) {
-    if ($expire == CACHE_TEMPORARY) {
-      $expire = 180;
-    }
-    
+  function set($key, $value, $expire = CACHE_PERMANENT, $headers = NULL) {    
     // Create new cache object.
     $cache = new stdClass;
     $cache->cid = $key;
@@ -63,6 +62,10 @@ class memcacheCache extends Cache {
     $cache->expire = $expire;
     $cache->headers = $headers;
     $cache->data = $value;
+    
+    if ($expire == CACHE_TEMPORARY || $expire == CACHE_PERMENANT) {
+      $set_expire = 0;  
+    }
     
     if (!empty($key)) {
       if ($this->settings['shared']) {
@@ -76,10 +79,10 @@ class memcacheCache extends Cache {
           }
 
           // Set key to 1 so we can keep track of the bin
-          $lookup[$this->key($key)] = 1;
+          $lookup[$this->key($key)] = $expire;
 
           // Attempt to store full key and value
-          if (!$this->memcache->set($this->key($key), $cache, $this->settings['compress'], $expire)) {
+          if (!$this->memcache->set($this->key($key), $cache, $this->settings['compress'], $set_expire)) {
             unset($lookup[$this->key($key)]);
             $return = FALSE;
           }
@@ -90,7 +93,7 @@ class memcacheCache extends Cache {
           }
 
           // Resave the lookup table (even on failure)
-          $this->memcache->set($this->lookup, $lookup, $this->settings['compress'], $expire);  
+          $this->memcache->set($this->lookup, $lookup, FALSE, 0);  
 
           // Remove lock.
           $this->unlock();
@@ -98,18 +101,41 @@ class memcacheCache extends Cache {
       }
       else {
         // Update memcache
-        return $this->memcache->set($this->key($key), $cache, $this->settings['compress'], $expire);
+        return $this->memcache->set($this->key($key), $cache, $this->settings['compress'], $set_expire);
       }
     }
   }
   
   function delete($key) {
     // Delete from static cache
-    parent::delete($this->key($key));
+    parent::flush();
     
-    // Remove from memcache.
-    return $this->memcache->delete($this->key($key));
-
+    if (substr($key, strlen($key) - 1, 1) == '*') {
+      $key = $this->key(substr($key, 0, strlen($key) - 1));
+      if ($this->settings['shared']) {
+        $lookup = $this->memcache->get($this->lookup);
+        if (!empty($lookup)) {
+          foreach ($lookup as $k => $v) {
+            if (substr($k, 0, strlen($key)) == $key) {
+              $this->memcache->delete($k);
+              unset($lookup[$k]);
+            }
+          }
+        }
+        if ($this->lock()) {
+          $this->memcache->set($this->lookup, $lookup, FALSE, 0); 
+          $this->unlock();
+        }
+      }
+      else {
+        return $this->flush();
+      }
+    }
+    else {
+      if (!empty($key)) {
+        return $this->memcache->delete($this->key($key));
+      }
+    }
   }
   
   function flush() {
@@ -130,14 +156,15 @@ class memcacheCache extends Cache {
         }
 
         // Cycle through keys and remove each entry from the cache
-        foreach ($lookup as $k => $v) {
-          if ($this->memcache->delete($k)) {
+        foreach ($lookup as $k => $expire) {
+          if ($expire != CACHE_PERMANENT && $expire < time()) {
+            $this->memcache->delete($k);
             unset($lookup[$k]);
           }
         }
 
         // Resave the lookup table (even on failure)
-        $this->memcache->set($this->lookup, $lookup, $this->settings['compress'], 0);
+        $this->memcache->set($this->lookup, $lookup, FALSE, 0);
 
         // Remove lock
         $this->unlock();
@@ -173,12 +200,30 @@ class memcacheCache extends Cache {
     foreach ($this->settings['servers'] as $server) {
       list($host, $port) = explode(':', $server);
       if (!$this->memcache->addServer($host, $port)) {
-        drupal_set_message("Unable to connect to memcache server $host:$port");
+        watchdog('cache', "Unable to connect to memcache server $host:$port", WATCHDOG_ERROR);
       }
     }
   }
   
   function close() {
     $this->memcache->close();
+  }
+  
+  function stats() {
+    $memcache_stats = $this->memcache->getStats();
+    $stats = array(
+      'uptime' => $memcache_stats['uptime'],
+      'bytes_used' => $memcache_stats['bytes'],
+      'bytes_total' => $memcache_stats['limit_maxbytes'],
+      'gets' => $memcache_stats['cmd_get'],
+      'sets' => $memcache_stats['cmd_set'],
+      'hits' => $memcache_stats['get_hits'],
+      'misses' => $memcache_stats['get_misses'],
+      'req_rate' => (($memcache_stats['cmd_get'] + $memcache_stats['cmd_set']) / $memcache_stats['uptime']),
+      'hit_rate' => ($memcache_stats['get_hits'] / $memcache_stats['uptime']),
+      'miss_rate' => ($memcache_stats['get_misses'] / $memcache_stats['uptime']),
+      'set_rate' => ($memcache_stats['cmd_set'] / $memcache_stats['uptime']),
+    );
+    return $stats;
   }
 }
